Fix problems found in a full review of the code
- Stop the crash when the screen size changes while compositing is off
- Give up adding windows when taking over the screen fails
- Free the overlay, colormap and selection when starting the compositor fails
- Build again when libepoxy is missing, and stop configure when it has no EGL
- Read the texture flip flag so windows are not drawn upside down
- Only use frame buffer blits when the driver really has them
- Apply the vsync setting to the window that is swapped, and keep --vblank
- Remove the dialog watch before the client goes away
- Check the number of desktops, icon sizes and other data sent by clients
- Read the session file without walking past its end
- Pick the right monitor when counting screens and finding clones
- Keep one sync alarm per window and stop re-arming a window that timed out
- Handle signals from the main loop instead of the signal handler
- Free what was leaked in menus, icons, tab window, settings dialogs and rc files
- Fix many small mistakes in upstream code found by the review
- Correct translations and stray spaces, markup and shortcuts in 33 catalogs
- Fix theme images whose header did not match the pixels
- Update README and COMPOSITOR to match what the code does

+++ src/placement.c
@@ -911,8 +911,8 @@ clientInitPosition (Client * c)
     if (place)
     {
         if ((screen_info->params->placement_ratio >= 100) ||
-            (100 * frameExtentWidth(c) * frameExtentHeight(c)) <
-                (screen_info->params->placement_ratio * full.width * full.height))
+            ((gint64) 100 * frameExtentWidth(c) * frameExtentHeight(c)) <
+                ((gint64) screen_info->params->placement_ratio * full.width * full.height))
         {
             if (screen_info->params->placement_mode == PLACE_MOUSE)
             {
